+++ lib/source/matrix/MatrixDot.cpp
@@ -17,8 +17,8 @@ namespace qlm
 			return Status::INVALID_DIMENSIONS;
 		}
 
-		unsigned int num_used_threads = pool.used_threads;
-		const unsigned int total_rows = rows;
+		size_t num_used_threads = pool.used_threads;
+		const size_t total_rows = rows;
 		num_used_threads = total_rows < num_used_threads ? total_rows : num_used_threads;
 
 		auto dot_mat = [](const int num_rows, const int start_r, const float* const __restrict  p_src1, 
@@ -41,21 +41,21 @@ namespace qlm
 			}
 		};
 		// divide the matrix among the threads
-		const unsigned int rows_per_thread = total_rows / num_used_threads;
-		const unsigned int thread_tail = total_rows % num_used_threads;
+		const size_t rows_per_thread = total_rows / num_used_threads;
+		const size_t thread_tail = total_rows % num_used_threads;
 		std::vector<std::future<void>> futures(num_used_threads);
 		// launch the threads
 		int next_row = 0;
 
 		#pragma omp unroll full
-		for (unsigned int i = 0; i < thread_tail; i++)
+		for (size_t i = 0; i < thread_tail; i++)
 		{
 			futures[i] = pool.Submit(dot_mat, rows_per_thread + 1, next_row, data, src.data, src.rows, src.columns, dst.data);
 			next_row += rows_per_thread + 1;
 		}
 
 		#pragma omp unroll full
-		for (unsigned int i = thread_tail; i < num_used_threads; i++)
+		for (size_t i = thread_tail; i < num_used_threads; i++)
 		{
 			futures[i] = pool.Submit(dot_mat, rows_per_thread, next_row, data, src.data, src.rows, src.columns, dst.data);
 			next_row += rows_per_thread;
@@ -63,7 +63,7 @@ namespace qlm
 
 		// wait for the threads to finish
 		#pragma omp unroll full
-		for (unsigned int i = 0; i < num_used_threads; i++)
+		for (size_t i = 0; i < num_used_threads; i++)
 		{
 			futures[i].wait();
 		}
