+++ lib/source/vector/common/VectorProc_2ArgScalar_Out.h
@@ -7,8 +7,8 @@
 
 namespace qlm
 {
-	template <void (*op)(const float, const unsigned int, float&, unsigned int&, float&, unsigned int&)>
-	inline Status Vector::VectorProc_2ArgScalar_Out(unsigned int& dst0, unsigned int& dst1, ThreadPool& pool) const
+	template <void (*op)(const float, const size_t, float&, size_t&, float&, size_t&)>
+	inline Status Vector::VectorProc_2ArgScalar_Out(size_t& dst0, size_t& dst1, ThreadPool& pool) const
 	{
 		if (pool.used_threads <= 0)
 		{
@@ -20,21 +20,21 @@ namespace qlm
 			return Status::INVALID_DIMENSIONS;
 		}
 
-		const unsigned int total_length = len;
-		const unsigned int num_used_threads = std::min(pool.used_threads, total_length);
+		const size_t total_length = len;
+		const size_t num_used_threads = std::min(pool.used_threads, total_length);
 
 
-		auto op_vec = [](const float* const __restrict  src, const unsigned int start_idx, const unsigned int size)
+		auto op_vec = [](const float* const __restrict  src, const size_t start_idx, const size_t size)
 		{
 			float dst0 = src[0];
-			unsigned int dst_idx0 = start_idx;
+			size_t dst_idx0 = start_idx;
 
 			float dst1 = src[0];
-			unsigned int dst_idx1 = start_idx;
+			size_t dst_idx1 = start_idx;
 
 #pragma loop( ivdep )
 #pragma omp simd
-			for (unsigned int i = 1; i < size; i++)
+			for (size_t i = 1; i < size; i++)
 			{
 				op(src[i], start_idx + i, dst0, dst_idx0, dst1, dst_idx1);
 			}
@@ -42,21 +42,21 @@ namespace qlm
 			return std::make_tuple(dst0, dst_idx0, dst1, dst_idx1);
 		};
 		// divide the vector among the threads
-		const unsigned int thread_length = total_length / num_used_threads;
-		const unsigned int thread_tail_length = total_length % num_used_threads;
-		std::vector<std::future<std::tuple<float, unsigned int, float, unsigned int>>> futures(num_used_threads);
+		const size_t thread_length = total_length / num_used_threads;
+		const size_t thread_tail_length = total_length % num_used_threads;
+		std::vector<std::future<std::tuple<float, size_t, float, size_t>>> futures(num_used_threads);
 		// launch the threads
-		unsigned int next_idx = 0;
+		size_t next_idx = 0;
 
 #pragma omp unroll full
-		for (unsigned int i = 0; i < thread_tail_length; i++)
+		for (size_t i = 0; i < thread_tail_length; i++)
 		{
 			futures[i] = pool.Submit(op_vec, &data[next_idx], next_idx, thread_length + 1);
 			next_idx += thread_length + 1;
 		}
 
 #pragma omp unroll full
-		for (unsigned int i = thread_tail_length; i < num_used_threads; i++)
+		for (size_t i = thread_tail_length; i < num_used_threads; i++)
 		{
 			futures[i] = pool.Submit(op_vec, &data[next_idx], next_idx, thread_length);
 			next_idx += thread_length;
@@ -71,7 +71,7 @@ namespace qlm
 
 		// wait for the threads to finish
 #pragma omp unroll full
-		for (unsigned int i = 1; i < num_used_threads; i++)
+		for (size_t i = 1; i < num_used_threads; i++)
 		{
 			auto result = futures[i].get();
 			op(std::get<0>(result), std::get<1>(result), dst0_val, dst0, dst1_val, dst1);
