+++ lib/source/vector/common/VectorProc_2Scalar_Out.h
@@ -24,17 +24,17 @@ namespace qlm
 			return Status::INVALID_DIMENSIONS;
 		}
 
-		const unsigned int total_length = len;
-		const unsigned int num_used_threads = std::min(pool.used_threads, total_length);
+		const size_t total_length = len;
+		const size_t num_used_threads = std::min(pool.used_threads, total_length);
 
 
-		auto op_vec = [](const float* const __restrict  src, const unsigned int size)
+		auto op_vec = [](const float* const __restrict  src, const size_t size)
 		{
 			float dst0 = src[0];
 			float dst1 = src[0];
 
 #pragma omp simd
-			for (unsigned int i = 1; i < size; i++)
+			for (size_t i = 1; i < size; i++)
 			{
 				op(src[i], dst0, dst1);
 			}
@@ -42,21 +42,21 @@ namespace qlm
 			return std::make_pair(dst0, dst1);
 		};
 		// divide the matrix among the threads
-		const unsigned int thread_length = total_length / num_used_threads;
-		const unsigned int thread_tail_length = total_length % num_used_threads;
+		const size_t thread_length = total_length / num_used_threads;
+		const size_t thread_tail_length = total_length % num_used_threads;
 		std::vector<std::future<std::pair<float, float>>> futures(num_used_threads);
 		// launch the threads
 		int next_idx = 0;
 
 #pragma omp unroll full
-		for (unsigned int i = 0; i < thread_tail_length; i++)
+		for (size_t i = 0; i < thread_tail_length; i++)
 		{
 			futures[i] = pool.Submit(op_vec, &data[next_idx], thread_length + 1);
 			next_idx += thread_length + 1;
 		}
 
 #pragma omp unroll full
-		for (unsigned int i = thread_tail_length; i < num_used_threads; i++)
+		for (size_t i = thread_tail_length; i < num_used_threads; i++)
 		{
 			futures[i] = pool.Submit(op_vec, &data[next_idx], thread_length);
 			next_idx += thread_length;
@@ -68,7 +68,7 @@ namespace qlm
 
 		// wait for the threads to finish
 #pragma omp unroll full
-		for (unsigned int i = 1; i < num_used_threads; i++)
+		for (size_t i = 1; i < num_used_threads; i++)
 		{
 			auto dst_th = futures[i].get();
 			op(dst_th.first, dst0, dst1);
